// src/pages/detail/detail.type.ts
export interface IProduct {
    id: number
    name: string
    alias: string
    price: number
    feature: boolean
    description: string
    size: string[]
    shortDescription: string
    quantity: number
    image: string
    categories: Category[]
    relatedProducts: RelatedProduct[]
}

interface Category {
    id: string
    category: string
}

interface RelatedProduct {
    id: number
    name: string
    alias: string
    feature: boolean
    price: number
    description: string
    shortDescription: string
    image: string
}